+++ subpackages/lib/source/odood/lib/deploy/config.d
@@ -96,6 +96,11 @@ struct DeployConfig {
     Path fail2ban_filter_path = Path("/", "etc", "fail2ban", "filter.d", "odoo-auth.conf");
     Path fail2ban_jail_path = Path("/", "etc", "fail2ban", "jail.d", "odoo-auth.conf");
 
+    bool letsencrypt_enable = false;
+    string letsencrypt_email = null;
+    Path letsencrypt_webroot = Path("/", "var", "/var/acme_challenge_webroot");
+    uint letsencrypt_rsa_key_size = 4096;
+
     Nullable!GitURL assembly_repo;
 
     /** Validate deploy config
@@ -173,6 +178,26 @@ struct DeployConfig {
             enforce!OdoodDeployException(
                 this.nginx.ssl_cert.exists(),
                 "SSL enabled, but provided SSL certificate (%s) does not exists!".format(this.nginx.ssl_cert));
+        }
+
+        if (this.letsencrypt_enable) {
+            enforce!OdoodDeployException(
+                !this.letsencrypt_email.empty,
+                "Let's Encrypt enabled, but lets encrypt email is not specified.");
+            enforce!OdoodDeployException(
+                !this.nginx.enable,
+                "Let's Encrypt enabled, but local nginx is not enabled.");
+            enforce!OdoodDeployException(
+                !this.nginx.ssl_on,
+                "Let's Encrypt enabled, but local nginx ssl is not enabled.");
+            enforce!OdoodDeployException(
+                !this.nginx.server_name,
+                "Let's Encrypt enabled, server-name not specified.");
+            Process("certbot")
+                .withArgs("--version")
+                .execute
+                .ensureOk!OdoodDeployException(
+                    "Let's Encrypt enabled, but certbot is not installed.");
         }
     }